feat: Use Ctrl + C to stop dev mode properly
