+++ Cargo.toml
@@ -31,3 +31,4 @@ lazy_static = "1.4.0"
 dofigen = "1"
 docker-compose-types = "0.2.2"
 open = "1"
+ctrlc = "3.2.3"
