+++ src/cli/dev.rs
@@ -1,10 +1,13 @@
 pub use clap::Args;
+use log::debug;
 
 use crate::cli::build::Build;
-use crate::cli::logs::{Logs, Service};
+use crate::cli::logs::Logs;
 use crate::cli::start::Start;
+use crate::cli::stop::Stop;
 use crate::cli::CliCommand;
 use crate::config::DEFAULT_CONFIG_FILE;
+use crate::docker_compose::Service;
 
 #[derive(Args)]
 pub struct Dev {
@@ -31,6 +34,11 @@ impl CliCommand for Dev {
         log::debug!("Run start");
         start.run();
 
+        ctrlc::set_handler(move || {
+            debug!("Stop asked");
+        })
+        .expect("Error setting Ctrl-C handler");
+
         let logs = Logs {
             services: vec![Service::App],
             follow: true,
@@ -38,5 +46,11 @@ impl CliCommand for Dev {
         };
         log::debug!("Run logs");
         logs.run();
+
+        let stop = Stop;
+        log::debug!("Run stop");
+        stop.run();
+
+        log::debug!("End of dev mode");
     }
 }
